Addressed code review comments
Addressed code review comments for the latest code review
